Fix perfs on click
